requirements.txt
networkx~=2.6.3
torch~=1.8.2+cu111
numpy~=1.20.3
PyYAML~=5.4.1
scikit-learn~=0.24.2
pecanpy~=2.0.0
numba~=0.54.0
python-louvain~=0.15
scipy~=1.7.1